fixing off-by-one in bug in selection_read:find_terms()
The old version was < when it should have been <= so that a match at the final
position of the string could be found. After confirming this I decided that I
was being too fancy for myself trying to optimize the case where the first
match is unsuccessful, and there is insufficient space for another match of the
current term, so I removed this optimization, which also fixes the bug.