Argv becomes Argm
I renamed almost all instances of argv to argm. The exceptions are:
 * isargvar() (it is short for "is argument variable")
 * int main(argc, argv), the compiler wouldn't complain, but the point is that
   Unix gives you all of your arguments as positional arguments, while within
   shell functions arguments will be named to the point that positional
   arguments will be named by the number of their position
 * Old_argv this is to convert internal argument maps to UNIX argv. Now that
   the internal object isn't called Argv, the "Old_" part is redundant, but
   this was confusing and involved enough of a change that I didn't want to
   tempt fate by making more changed code look like code needing change.

This caused most of the copyright dates to change, most of them for the first
time since 2007 or 2008.

max_nesting_v became static, which doesn't change anything because there was
only one instance of the class ever used.

var_map became parent_map and stopped being static; i just explicitly
initialize them to all be the same.

** Note that these Argm objects are still vectors. I wanted to change the name
   first and having done so there are so many changes in the working copy that
   I didn't want to do another complex change without a commit.