making use of local variables in init scripts
ntimes which really needed parameters got them, but most other functions were
left as they were, especially after i found that i was giving a variable number
of arguments to rwsh.autofunction.

This let me remove the TEMP1 global, which was reused by every function that
had need of a variable, in a way that seemed guaranteed to fail in some
situation.

I also increased MAX_NESTING to 10 in .init, because at 4 i couldn't run .init
from a shell, and because I didn't see a reason to ensure that this was at an
absolute minimum.